+++ src/MetaheuristicsPlatform/Algorithms/Constraints/JoinesHouckPenaltyGa/JoinesHouckPenaltyGaOptimizerReferences.cs
@@ -1,0 +1,14 @@
+using MetaheuristicsPlatform.Classification;
+
+namespace MetaheuristicsPlatform.Algorithms.Constraints.JoinesHouckPenaltyGa;
+
+public static class JoinesHouckPenaltyGaOptimizerReferences
+{
+    public static ScientificReference Primary { get; } =
+        new(
+            "Joines & Houck",
+            1994,
+            "On the use of non-stationary penalty functions to solve nonlinear constrained optimization problems with GA's",
+            "Proceedings of the First IEEE Conference on Evolutionary Computation",
+            "10.1109/ICEC.1994.349995");
+}